Add extension methods for various types
This commit introduces multiple extension methods for various types that work to assist in object conversion, integer manipulation, checking the characteristics/status of a Guid, enum conversion methods, and a name equality comparer. Useful methods include ones to determine if a value can be assigned to a PropertyInfo, if a Guid is empty, and various mathematical manipulations on integers.